+++ backend/index.js
@@ -1,6 +1,8 @@
 const express = require('express')
 const mongoose = require('mongoose')
-const authRouter = require('./authRouter')
+const authRouter = require('./routers/authRouter')
+
+
 const PORT = process.env.PORT || 5000
 
 const app = express()
