+++ backend/routers/authRouter.js
@@ -1,6 +1,6 @@
 const Router = require('express')
 const router = new Router()
-const controller = require('./authController')
+const controller = require('../Controllers/authController')
 
 router.post('/registration', controller.registration)
 router.post('/login', controller.login)
